feat(resources): update to use mini cards
