+++ src/components/ComponentList/ComponentList.js
@@ -98,77 +98,6 @@ export class ComponentList extends React.Component {
 }
 
 /**
- * Returns a table for a single component status
- * @param {string} name Name of the component
- * @param {string} type Type of component (ui|layout)
- * @returns {*} Component status table
- * @constructor
- */
-export const ComponentStatus = ({ name, type }) => (
-  <div className="bx--row component-list">
-    <div className="bx--col-lg-12 bx--no-gutter">
-      <table className="page-table">
-        <thead>
-          <tr>
-            <th>Language</th>
-            <th>Status</th>
-            <th>Sandbox</th>
-          </tr>
-        </thead>
-        <tbody>
-          <tr key={`${name}-react`}>
-            <td>React</td>
-            <td>
-              {Object.keys(components[type][name].react)
-                .filter((key) => components[type][name].react[key])
-                .map((key) => (
-                  <React.Fragment key={key}>{_tags[key]}</React.Fragment>
-                ))}
-            </td>
-            <td>
-              {components[type][name].storybook.react !== "" ? (
-                <a
-                  href={components[type][name].storybook.react}
-                  target="_blank"
-                  rel="noreferrer"
-                >
-                  View Storybook
-                </a>
-              ) : (
-                "Coming soon"
-              )}
-            </td>
-          </tr>
-          <tr key={`${name}-webcomponents`}>
-            <td>Web Components</td>
-            <td>
-              {Object.keys(components[type][name].webcomponents)
-                .filter((key) => components[type][name].webcomponents[key])
-                .map((key) => (
-                  <React.Fragment key={key}>{_tags[key]}</React.Fragment>
-                ))}
-            </td>
-            <td>
-              {components[type][name].storybook.webcomponents !== "" ? (
-                <a
-                  href={components[type][name].storybook.webcomponents}
-                  target="_blank"
-                  rel="noreferrer"
-                >
-                  View Storybook
-                </a>
-              ) : (
-                "Coming soon"
-              )}
-            </td>
-          </tr>
-        </tbody>
-      </table>
-    </div>
-  </div>
-);
-
-/**
  * Returns the table with the list of tabs and their definitions
  * @returns {*} Tag key table
  * @constructor
