+++ src/components/ResourceLinks/index.js
@@ -1,0 +1,3 @@
+import { ResourceLinks } from './ResourceLinks';
+
+export default ResourceLinks;
