--- src/components/SpecLinks/index.js
@@ -1,3 +1,0 @@
-import { SpecLinks } from './SpecLinks';
-
-export default SpecLinks;
