+++ src/pages/components/button-group.mdx
@@ -5,6 +5,7 @@ description: The Button group component is to be utilized within IBM.com for gro
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Button group" type="ui" />
 
@@ -17,11 +18,11 @@ import ComponentFooter from "components/ComponentFooter";
 <AnchorLinks>
 
 <AnchorLink>Content guidance</AnchorLink>
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Button group" type="ui" />
 
 ## Content guidance
