+++ src/pages/components/callout-with-media.mdx
@@ -5,6 +5,7 @@ description: Callout with media is used to feature a high value media asset such
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Callout with media" type="layout" />
 
@@ -12,11 +13,11 @@ import ComponentFooter from "components/ComponentFooter";
 
 <AnchorLink>Default</AnchorLink>
 <AnchorLink>Content guidance</AnchorLink>
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Callout with media" type="layout" />
 
 ## Default
