+++ src/pages/components/card-in-card.mdx
@@ -5,6 +5,7 @@ description: Card in card offers a large media moment, while keeping harmony in 
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Card in card" type="ui" />
 
@@ -17,11 +18,11 @@ import ComponentFooter from "components/ComponentFooter";
 <AnchorLinks>
 
 <AnchorLink>Default</AnchorLink>
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Card in card" type="ui" />
 
 ## Default
