+++ src/pages/components/card-section.mdx
@@ -5,6 +5,7 @@ description: Card section is a collection of cards presented in a full-width sec
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Card section" type="layout" />
 
@@ -13,11 +14,11 @@ import ComponentFooter from "components/ComponentFooter";
 <AnchorLink>Default</AnchorLink>
 <AnchorLink>Variations</AnchorLink>
 <AnchorLink>Content guidance</AnchorLink>
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Card section" type="layout" />
 
 ## Default
