+++ src/pages/components/carousel.mdx
@@ -5,6 +5,7 @@ description: Carousel is a flexible component that allows users to explore a lar
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Carousel" type="ui" />
 
@@ -16,10 +17,10 @@ import ComponentFooter from "components/ComponentFooter";
 
 <AnchorLinks>
 
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Carousel" type="ui" />
 
 <ComponentFooter name="Carousel" type="ui" />
