+++ src/pages/components/content-block-cards.mdx
@@ -5,6 +5,7 @@ description: Content block cards is used to present small self-contained pieces 
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Content block cards" type="layout" />
 
@@ -13,11 +14,11 @@ import ComponentFooter from "components/ComponentFooter";
 <AnchorLink>Default</AnchorLink>
 <AnchorLink>Variations</AnchorLink>
 <AnchorLink>Content guidance</AnchorLink>
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Content block cards" type="layout" />
 
 ## Default
