+++ src/pages/components/content-block-horizontal.mdx
@@ -5,17 +5,18 @@ description: Content block horizontal can be used to present important areas of 
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Content block horizontal" type="layout" />
 
 <AnchorLinks>
 
 <AnchorLink>Content guidance</AnchorLink>
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Content block horizontal" type="layout" />
 
 ## Content guidance
