+++ src/pages/components/content-block-media.mdx
@@ -5,6 +5,7 @@ description: Content block media is used to present information with images in a
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Content block media" type="layout" />
 
@@ -13,11 +14,11 @@ import ComponentFooter from "components/ComponentFooter";
 <AnchorLink>Default</AnchorLink>
 <AnchorLink>Variations</AnchorLink>
 <AnchorLink>Content guidance</AnchorLink>
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Content block media" type="layout" />
 
 ## Default
