+++ src/pages/components/content-group-banner.mdx
@@ -5,6 +5,7 @@ description: Content group banner is designed as a subtle callout, intended for 
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Content group banner" type="layout" />
 
@@ -17,11 +18,11 @@ import ComponentFooter from "components/ComponentFooter";
 <AnchorLinks>
 
 <AnchorLink>Default</AnchorLink>
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Content group banner" type="layout" />
 
 ## Default
