+++ src/pages/components/content-item-horizontal.mdx
@@ -5,6 +5,7 @@ description: The content item horizontal component displays information in a hor
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Content item horizontal" type="layout" />
 
@@ -18,11 +19,11 @@ import ComponentFooter from "components/ComponentFooter";
 
 <AnchorLink>Default</AnchorLink>
 <AnchorLink>Variations</AnchorLink>
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Content item horizontal" type="layout" />
 
 ## Default
