+++ src/pages/components/cta.mdx
@@ -9,6 +9,7 @@ import ArrowDown20 from "@carbon/icons-react/es/arrow--down/20";
 import ArrowLeft20 from "@carbon/icons-react/es/arrow--left/20";
 import Launch20 from "@carbon/icons-react/es/launch/20";
 import PlayOutline20 from "@carbon/icons-react/es/play--outline/20";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="CTA" type="ui" />
 
@@ -16,11 +17,11 @@ import PlayOutline20 from "@carbon/icons-react/es/play--outline/20";
 
 <AnchorLink>Types</AnchorLink>
 <AnchorLink>Styles</AnchorLink>
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="CTA" type="ui" />
 
 ## Types
