+++ src/pages/components/dotcom-shell.mdx
@@ -5,6 +5,7 @@ description: The Dotcom shell component includes the Masthead, and Footer compon
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Dotcom shell" type="ui" />
 
@@ -16,9 +17,10 @@ import ComponentFooter from "components/ComponentFooter";
 
 <AnchorLinks>
 
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Dotcom shell" type="ui" />
 
 <ComponentFooter name="Dotcom shell" type="ui" />
