+++ src/pages/components/feature-section.mdx
@@ -5,6 +5,7 @@ description: Feature section is a dedicated section for a single story.
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Feature section" type="layout" />
 
@@ -17,12 +18,11 @@ Additional usage guidelines will be available soon.
 <AnchorLinks>
 
 <AnchorLink>Default</AnchorLink>
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
 
+<ResourceLinks name="Feature section" type="layout" />
 
 ## Default
