+++ src/pages/components/filter-panel.mdx
@@ -5,6 +5,7 @@ description: The filter panel is a component that can be customized to provide c
 
 import ComponentDescription from 'components/ComponentDescription';
 import ComponentFooter from 'components/ComponentFooter';
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Filter panel" type="ui" />
 
@@ -15,12 +16,11 @@ import ComponentFooter from 'components/ComponentFooter';
 <AnchorLink>Selection methods</AnchorLink>
 <AnchorLink>Behaviors</AnchorLink>
 <AnchorLink>Recommendations</AnchorLink>
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
 
+<ResourceLinks name="Filter panel" type="ui" />
 
 ## Default
