+++ src/pages/components/footer.mdx
@@ -5,6 +5,7 @@ description: The Footer component page
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Footer" type="ui" />
 
@@ -14,11 +15,11 @@ import ComponentFooter from "components/ComponentFooter";
 <AnchorLink>Variations</AnchorLink>
 <AnchorLink>Anatomy</AnchorLink>
 <AnchorLink>Behavior</AnchorLink>
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Footer" type="ui" />
 
 ## Default
