+++ src/pages/components/horizontal-rule.mdx
@@ -5,6 +5,7 @@ description: The Horizontal rule can be used as a graphical element to add struc
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Horizontal rule" type="ui" />
 
@@ -12,11 +13,11 @@ import ComponentFooter from "components/ComponentFooter";
 
 <AnchorLink>Default</AnchorLink>
 <AnchorLink>Variations</AnchorLink>
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Horizontal rule" type="ui" />
 
 ## Default
