+++ src/pages/components/image.mdx
@@ -5,6 +5,7 @@ description: The Image component is used as the primary way of embedding images 
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Image" type="ui" />
 
@@ -16,9 +17,10 @@ import ComponentFooter from "components/ComponentFooter";
 
 <AnchorLinks>
 
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Image" type="ui" />
 
 <ComponentFooter name="Image" type="ui" />
