+++ src/pages/components/layout.mdx
@@ -5,6 +5,7 @@ description: The Layout component is to be utilized within IBM.com for various a
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Layout" type="layout" />
 
@@ -16,9 +17,10 @@ import ComponentFooter from "components/ComponentFooter";
 
 <AnchorLinks>
 
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Layout" designLink={false} type="layout" />
 
 <ComponentFooter name="Layout" designLink={false} type="layout" />
