+++ src/pages/components/lead-space-block.mdx
@@ -5,6 +5,7 @@ description: Lead space &mdash; block is an alternative to the Lead space, enabl
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Lead space block" type="layout" />
 
@@ -12,11 +13,11 @@ import ComponentFooter from "components/ComponentFooter";
 
 <AnchorLink>Default</AnchorLink>
 <AnchorLink>Content guidance</AnchorLink>
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Lead space block" type="layout" />
 
 ## Default
