+++ src/pages/components/lead-space-search.mdx
@@ -5,6 +5,7 @@ description: Lead space search provides a fast route to information discovery by
 
 import ComponentDescription from 'components/ComponentDescription';
 import ComponentFooter from 'components/ComponentFooter';
+import ResourceLinks from "components/ResourceLinks";
 
 import ComponentFeedback from 'components/ComponentFeedback';
 
@@ -16,11 +17,10 @@ import ComponentFeedback from 'components/ComponentFeedback';
 <AnchorLink>Options</AnchorLink>
 <AnchorLink>Anatomy</AnchorLink>
 <AnchorLink>Behaviors</AnchorLink>
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 </AnchorLinks>
 
+<ResourceLinks name="Lead space search" type="layout" />
 
 ## Overview
