+++ src/pages/components/leadspace.mdx
@@ -5,6 +5,7 @@ description: A Lead space is positioned at the top of a web page and serves as t
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Lead space" type="layout" />
 
@@ -16,12 +17,12 @@ import ComponentFooter from "components/ComponentFooter";
 <AnchorLink>Usage guidance</AnchorLink>
 <AnchorLink>Color themes</AnchorLink>
 <AnchorLink>Type styles</AnchorLink>
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Resources</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Lead space" type="layout" />
 
 ## Lead space left-aligned
