+++ src/pages/components/leaving-ibm.mdx
@@ -5,17 +5,18 @@ description: The Leaving IBM displays as an overlay when users click an external
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Leaving IBM" type="ui" />
 
 <AnchorLinks>
 
 <AnchorLink>Default</AnchorLink>
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Leaving IBM" type="ui" />
 
 ## Default
