+++ src/pages/components/lightbox-media-viewer.mdx
@@ -5,17 +5,18 @@ description: The Lightbox media viewer allows the user to view an image or video
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Lightbox media viewer" type="ui" />
 
 <AnchorLinks>
 
 <AnchorLink>Default</AnchorLink>
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Lightbox media viewer" type="ui" />
 
 ## Default
