+++ src/pages/components/link-with-icon.mdx
@@ -5,17 +5,18 @@ description: The Link with icon is primarily used as a navigational element with
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Link with icon" type="ui" />
 
 <AnchorLinks>
 
 <AnchorLink>Default</AnchorLink>
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Link with icon" type="ui" />
 
 ## Default
 Link with icon should not be used within a paragraph. The icon can be placed on the left or right side of the text and is vertically centered.
