+++ src/pages/components/localemodal.mdx
@@ -5,17 +5,18 @@ description: The Locale modal allows users to change geographic regions and lang
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Locale modal" type="ui" />
 
 <AnchorLinks>
 
 <AnchorLink>Default</AnchorLink>
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Locale modal" type="ui" />
 
 ## Default
