+++ src/pages/components/masthead.mdx
@@ -5,6 +5,7 @@ description: The Masthead component page
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Masthead" type="ui" />
 
@@ -12,11 +13,11 @@ import ComponentFooter from "components/ComponentFooter";
 
 <AnchorLink>Masthead with Level 0</AnchorLink>
 <AnchorLink>Masthead with Level 1</AnchorLink>
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Masthead" type="ui" />
 
 ## Masthead with Level 0
 The Masthead with Level 0 (L0) refers to the first position of navigational menus on a page, it houses global elements and navigation labels. L is abbreviated for the "Level" starting at base 0, the Level refers to the hierarchy of the navigation on the page.
