+++ src/pages/components/pictogram-item.mdx
@@ -5,6 +5,7 @@ description: The Pictogram item is used to communicate messages at a glance, aff
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Pictogram item" type="layout" />
 
@@ -16,10 +17,10 @@ import ComponentFooter from "components/ComponentFooter";
 
 <AnchorLinks>
 
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Pictogram item" type="layout" />
 
 <ComponentFooter name="Pictogram item" type="layout" />
