+++ src/pages/components/quote.mdx
@@ -5,6 +5,7 @@ description: The Quote component is used to highlight an impactful client statem
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Quote" type="ui" />
 
@@ -16,10 +17,10 @@ import ComponentFooter from "components/ComponentFooter";
 
 <AnchorLinks>
 
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Quote" type="ui" />
 
 <ComponentFooter name="Quote" type="ui" />
