+++ src/pages/components/tabs-extended.mdx
@@ -5,6 +5,7 @@ description: Tabs extended combines tabs and accordion to organize editorial con
 
 import ComponentDescription from "components/ComponentDescription";
 import ComponentFooter from "components/ComponentFooter";
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Tabs extended" type="ui" />
 
@@ -17,11 +18,11 @@ import ComponentFooter from "components/ComponentFooter";
 <AnchorLinks>
 
 <AnchorLink>Default</AnchorLink>
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Tabs extended" type="ui" />
 
 ## Default
