+++ src/pages/components/tag-link.mdx
@@ -3,9 +3,9 @@ title: Tag link
 description: Tag link is a clickable tag that links the topic to another page for users to find more information about the topic.
 ---
 
-import { ComponentStatus } from 'components/ComponentList';
 import ComponentDescription from 'components/ComponentDescription';
 import ComponentFeedback from 'components/ComponentFeedback';
+import ResourceLinks from "components/ResourceLinks";
 
 <ComponentDescription name="Tag link" type="ui" />
 
@@ -16,12 +16,13 @@ import ComponentFeedback from 'components/ComponentFeedback';
 </InlineNotification>
 
 <AnchorLinks>
+
 <AnchorLink>Default</AnchorLink>
-<AnchorLink>Design and functional specifications</AnchorLink>
-<AnchorLink>Development documentation</AnchorLink>
 <AnchorLink>Feedback</AnchorLink>
 
 </AnchorLinks>
+
+<ResourceLinks name="Tag link" designLink={"https://ibm.box.com/s/y5dx0hxzcsn6i6xv6i1vwlsec300gbm5"} type="ui" />
 
 ## Default
 
@@ -34,15 +35,6 @@ Tag link is a clickable tag that links the topic to another page for users to fi
 
 </Column>
 </Row>
-
-## Design and functional specifications
-
-For more details on Tag link, see [design specs](https://ibm.box.com/s/y5dx0hxzcsn6i6xv6i1vwlsec300gbm5) and see [functional specs](https://github.com/carbon-design-system/carbon-for-ibm-dotcom-website/wiki/Tag-link).
-
-
-## Development documentation
-
-<ComponentStatus name="Tag link" type="ui" />
 
 ## Feedback
